Moving functional sections to single classes
