--- Helpers/app.config
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="AnotherRoadUpdateTool.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
-        </sectionGroup>
-    </configSections>
-    <userSettings>
-        <AnotherRoadUpdateTool.Properties.Settings>
-            <setting name="AdjustAreas" serializeAs="String">
-                <value>False</value>
-            </setting>
-            <setting name="AdjustMoney" serializeAs="String">
-              <value>False</value>
-            </setting>
-            <setting name="AllRoads" serializeAs="String">
-                <value>False</value>
-            </setting>
-            <setting name="Chirper" serializeAs="String">
-              <value>True</value>
-            </setting>
-            <setting name="MaxAreas" serializeAs="String">
-                <value>25</value>
-            </setting>
-            <setting name="StartMoney" serializeAs="String">
-                <value>150000</value>
-            </setting>
-        </AnotherRoadUpdateTool.Properties.Settings>
-    </userSettings>
-</configuration>
